Recycler.Update.allowSynchronousUpdate
This flag tells if an update *can* be done synchronously, by sending notifications to the adapter directly inside the update call. Defaults to false.

This way a user can opt-in into this behavior which, in certain usages, might cause issues (Recycler does not like notifications when there's a pending relayout, or the recycler is scrolling).